chore(int64): mail_selector, sort keys, and remaining product atoi (#1402)
- Widen mail_selector low/high/days to int64_t (engine + module headers).
- mail_mod message-list parse: atol → mux_atoi64.
- color_ops sort: parse_i64 via strtoll (regenerate color_ops.c).
- DBT TINYMUX_DBT_PAD env: strtoll instead of atoi.
- Hydra proxy session timestamps: std::atoll (time_t-width).

Product softcode/server path is free of C atoi/atol and int sinks after
mux_atoi64. Offline convert/lexers and mux_atol unit tests stay as-is.

+++ mux/include/mail.h
@@ -64,11 +64,13 @@ struct mail
 
 struct mail_selector
 {
-    int       low;
-    int       high;
+    // Message index / age filters: full-width integers (#1402).
+    //
+    int64_t   low;
+    int64_t   high;
     mail_flag flags;
     dbref     player;
-    int       days;
+    int64_t   days;
     int       day_comp;
 };
